api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.37
asyncpg==0.30.0
pydantic-settings==2.7.1
httpx==0.28.1
